+++ CHANGELOG.md
@@ -9,6 +9,7 @@ with respect to its command line interface and HTTP interface.
 ## [Unreleased](//github.com/opentable/sous/compare/0.5.86...HEAD)
 ### Changed
 * Client: `sous init` now requires -kind flag which is either 'scheduled' or 'http-service'
+* Client: error is returned if manifest set is sent a different source location
 
 ## [0.5.86](//github.com/opentable/sous/compare/0.5.85...0.5.86)
 ### Added