+++ CHANGELOG.md
@@ -4,11 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
-with respect to its command line interface and HTTP interface.
+with respect to its command line interface and HTTP interface
 
 ## [Unreleased](//github.com/opentable/sous/compare/0.5.87...HEAD)
 ### Added
 * Both: Sous commands that communicate with the server add a request ID header to trace the request.
+
+### Changed
+* Client: error is returned if manifest set is sent a different source location
+
 
 ## [0.5.86](//github.com/opentable/sous/compare/0.5.85...0.5.86)
 ### Added