+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface
 
-## [Unreleased](//github.com/opentable/sous/compare/0.5.87...HEAD)
+## [0.5.88](//github.com/opentable/sous/compare/0.5.87...0.5.88)
 ### Added
 * Both: Sous commands that communicate with the server add a request ID header to trace the request.